bazel: make `fastbuild` the default `-c` again; never strip for `dev` builds
This is a partial revert of `60e8bcec61269c6648dc40c8094fbf303b8a02aa`.

In that commit I made a change to make un-stripped binaries the default
by making `dbg` the default `--compilation_mode`. This had unexpected
consequences as this actually disables inlining and optimization,
thereby breaking some tests, and overall it is surprising that a Go
binary built by Bazel would be un-optimized by default since this is
the opposite of `go build`'s default behavior.

Instead, we make `fastbuild` the default (again), and to solve the
problem of binaries being unstripped we set `--strip never` for
`dev` builds. (`dev doctor` makes you clarify that your build is
a `dev` build on setup so no one can miss this.)

Now we have the following behavior for each `compilation_mode`:

* `dbg`: disable optimizations and inlining.
* `fastbuild` and `opt`: will produce the same binary, except
   `fastbuild` defaults to stripping in non-`dev` configurations.

If you want an un-optimized binary for some reason, it still works to
build with `-c dbg`.

Epic: CRDB-17171
Release note: None